requirements.txt
Django==1.11.17
Markdown==2.5
PyJWT==1.4.2
amqp==1.4.9
setuptools==32.1.2
anyjson==0.3.3
argparse==1.2.1
billiard==3.3.0.23
celery==3.1.23
dj-database-url==0.5
django-celery==3.2.2
django-celery-with-redis==3.0
django-compressor==2.2
django-filter==1.1
django-jsonify==0.3.0
django-polymorphic==1.3.1
django-redis==1.4.5
django-smtp-ssl==1.0
django-yamlfield==1.0.1
djangorestframework==3.6.4
# Elasticsearch 5.x
elasticsearch==5.2.0
elasticsearch-dsl==5.3.0
gunicorn==18.0
gevent==1.0.1
hiredis==0.1.1
httplib2==0.7.7
icalendar==3.11
kombu==3.0.34
mock==1.0.1
psycogreen==1.0
psycopg2==2.5.4
pytimeparse==1.1.6
pytz==2014.10
redis==2.9.0
requests==2.21.0
six==1.12.0
twilio==6.16.1
wsgiref==0.1.2
python-dateutil==2.1
django-auth-ldap==1.2.6
PyYAML==3.11
social-auth-core==2.0.0
social-auth-app-django==3.1.0
boto==2.38.0
rollbar==0.14.6
unittest-xml-reporting==2.1.0
PyMySQL==0.8.0
jsondiff==1.1.2
django-timezone-field==3.0
django-reversion==1.8.7
